requirements.txt
Django==3.0.6
django-crispy-forms==1.9.1
django-ckeditor==5.9.0